% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-consensus_inner.r
\name{consensus_inner}
\alias{consensus_inner}
\title{consensus_inner}
\usage{
consensus_inner(direction = c("right", "left"), consensus, haps, riskhaps,
  site.start, site.end, groups, max.leave, verbose)
}
\arguments{
\item{direction}{Which direction to move from site of interest. Options are "right" (genomic positions higher than site of interest) and "left" (genomic positions lower than site of interest).}

\item{consensus}{Skeleton data frame to which consensus haplotype segments will be added}

\item{haps}{Matrix of phased haplotypes}

\item{riskhaps}{Vector of identifiers for risk haplotypes}

\item{site.start}{Genomic position at start of site of interest}

\item{site.end}{Genomic position at end of site of interest}

\item{groups}{Internal variable tracking which subjects are in the current haplotype group}

\item{max.leave}{Maximum number of haplotypes that may leave a haplotype group without starting their own new consensus haplotype segment}

\item{verbose}{Logical indicating whether progress updates should be printed. Default FALSE.}
}
\value{
List containing data frame of consensus haplotype segments and vector of split positions.
}
\description{
Finds the consensus risk haplotype shared between case subjects
(main body of algorithm for get.consensus function).
}
\keyword{internal}
