% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a-consensus_summary.r
\name{consensus_summary}
\alias{consensus_summary}
\title{consensus_summary}
\usage{
consensus_summary(consensus)
}
\arguments{
\item{consensus}{Data frame with all consensus haplotype segments for each case subject (output from get.consensus)}
}
\value{
Data frame containing risk haplotype identifier and overall start/end of consensus haplotype
}
\description{
Summarizes overall start and end position of consensus risk haplotype
for each case individual (on their risk haplotype).
}
