% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b-end_haplotypes.r
\name{end_haplotypes}
\alias{end_haplotypes}
\title{end_haplotypes}
\usage{
end_haplotypes(direction, consensus, riskhaps, haps, end.pos, ends, groups,
  this.hapid)
}
\arguments{
\item{direction}{Which direction to move from site of interest. Options are "right" (genomic positions higher than site of interest) and "left" (genomic positions lower than site of interest).}

\item{consensus}{Data frame containing consensus haplotype segments.}

\item{riskhaps}{Vector of identifiers for risk haplotypes}

\item{haps}{Matrix of phased haplotypes}

\item{end.pos}{Current end position of consensus haplotype segment}

\item{ends}{Vector storing end.pos for all consensus haplotype segments}

\item{groups}{Internal variable tracking which subjects are in the current haplotype group}

\item{this.hapid}{Consensus haplotype segment currently under consideration}
}
\value{
Consensus, groups, end.pos, ends
}
\description{
Ends consensus risk haplotype for particular subjects (when fewer than
max.leave differ from majority)
}
\keyword{internal}
