% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c-end_pair.r
\name{end_pair}
\alias{end_pair}
\title{end_pair}
\usage{
end_pair(direction = c("right", "left"), consensus, haps, this.hapid, groups,
  ends)
}
\arguments{
\item{direction}{Which direction to move from site of interest. Options are "right" (genomic positions higher than site of interest) and "left" (genomic positions lower than site of interest).}

\item{consensus}{Skeleton data frame to which consensus haplotype segments will be added}

\item{haps}{Matrix of phased haplotypes}

\item{this.hapid}{Consensus haplotype segment currently under consideration}

\item{groups}{Internal variable tracking which subjects are in the current haplotype group}

\item{ends}{Vector storing end.pos for all consensus haplotype segments}
}
\value{
Consensus, groups, ends
}
\description{
Ends consensus risk haplotype for a pair of subjects when they are the
last remaining pair in a consensus haplotype segment
}
\keyword{internal}
