% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-toy-data.r
\name{generate_toy_data}
\alias{generate_toy_data}
\title{Generates simulated data for testing CHap functions.}
\usage{
generate_toy_data(seed = 1000, n.markers = 50, n.subj = 10,
  marker.spacing = 0.1, num.splits = 5)
}
\arguments{
\item{seed}{Random seed for data generation; default 1.}

\item{n.markers}{Number of markers on simulated "chromosome"; default 50.}

\item{n.subj}{Number of subjects; default 10, all case subjects.}

\item{marker.spacing}{Distance between markers; default 0.1 (as if cM positions).}

\item{num.splits}{Number of times the consensus haplotype splits on each side}
}
\value{
Produces 3 files in the current directory: sim-case-ids.txt, sim-vcf.txt, and sim-ibs.txt. Returns R objects "haps" (the haplotype data frame for input into consensus) and "riskhaps" (the vector of true risk haplotypes)
}
\description{
Generates simulated data for testing CHap functions.
}
\keyword{internal}
