% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-get_consensus.r
\name{get_consensus}
\alias{get_consensus}
\title{get_consensus}
\usage{
get_consensus(haps, riskhaps, site.start, site.end, verbose = F)
}
\arguments{
\item{haps}{Matrix of phased haplotypes. Rows are markers, columns are copies of the chromosome (2 columns per subject). Must have a column called POS containing the genetic positions of the markers. Other column names should have the format SUBJ1.1 and SUBJ1.2 for the two haplotypes of SUBJ1.}

\item{riskhaps}{Vector of risk haplotype identifiers (one per case subject).}

\item{site.start}{Genomic position at start of site of interest.}

\item{site.end}{Genomic position at end of site of interest.}

\item{verbose}{Logical indicating whether progress updates should be printed. Default FALSE.}
}
\value{
List containing data frame of consensus haplotype segments and vector of split positions.
}
\description{
Finds the consensus risk haplotype shared between case subjects
}
