% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{haps}
\alias{haps}
\title{Simulated haplotype matrix in format for input into get.consensus().}
\format{A matrix with 50 rows and 21 columns:
\describe{
  \item{POS}{Genomic position of marker}
  \item{SUBJID1.1}{Data for subject 1, haplotype 1.}
  ...
}}
\usage{
haps
}
\description{
Simulated haplotype matrix in format for input into get.consensus().
}
\keyword{datasets}
