% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a-plot_heatmap.r
\name{ibs_heatmap}
\alias{ibs_heatmap}
\title{plot_heatmap}
\usage{
ibs_heatmap(consensus, groups = NULL, group.lines = FALSE,
  graphmin = NULL, graphmax = NULL, main = NULL, cx = 0.75)
}
\arguments{
\item{consensus}{Data frame containing consensus haplotype segments (columns are riskhap ID, segment start, segment end, and consensus haplotype segment ID. Output from get.consensus().}

\item{groups}{Vector indicating group membership. Default NULL; if included, subjects will be ordered by group (i.e., subjects in same group will be plotted next to each other.)}

\item{group.lines}{Logical indicating whether groups should be marked on plot. Default FALSE; if TRUE, gray lines will demarcate groups.}

\item{graphmin}{Minimum IBS value for plot. Anything lower (but nonzero) will be increased to this value for the plot. Default NULL.}

\item{graphmax}{Maximum IBS value for plot. Anything higher willl be truncated to this value for the plot. Default NULL.}

\item{main}{Title of plot}

\item{cx}{Multiplier for font size for subject labels, graph title. Default 0.75.}
}
\value{
Creates IBS heat map.
}
\description{
Generates heatmap to visualize pairwise IBS sharing between case subjects
at the site of interest
}
