% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5b-ibs_matrix.r
\name{ibs_matrix}
\alias{ibs_matrix}
\title{ibs_matrix}
\usage{
ibs_matrix(consensus, groups = NULL)
}
\arguments{
\item{consensus}{Data frame containing consensus haplotype segments (columns are riskhap ID, segment start, segment end, and consensus haplotype segment ID. Output from get.consensus().}

\item{groups}{Vector indicating group membership.}
}
\value{
Matrix of pairwise IBS sharing between case individuals across site of interest.
}
\description{
Generates matrix of pairwise IBS sharing between case subjects
at the site of interest
}
