% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-id_riskhap.r
\name{id_riskhap}
\alias{id_riskhap}
\title{id_riskhap}
\usage{
id_riskhap(ibs.file, case.id.file, site.start, site.end)
}
\arguments{
\item{ibs.file}{Path to the input IBS file}

\item{case.id.file}{Path to the input case ID file}

\item{site.start}{Genomic position at start of site of interest}

\item{site.end}{Genomic position at end of site of interest}
}
\value{
A vector of risk haplotype IDs
}
\description{
Identify which haplotype is the risk haplotype for each case individual.
Note that we assume only one haplotype carries the feature at the site
of interest.
}
\details{
This function loads an IBS file and a case ID file, and also requires the
start and end locations of the site of interest to be specified.

The IBS file should have eight columns: ID1, Hap1, ID2, Hap2, Chrom,
Start, End, Length. It may be generated using Brian Browning's ibs.jar
utility, located at [].

The case ID file should be a text file with a single column containing all
of the case subject identifiers.

site.start and site.end are the genomic positions that begin and end the
site of interest, in the same units as those used in the IBS file and the
vcf/haplotype file.

Of the two haplotypes for a given case, the one with the longest median IBS
length at the site of interest (shared with other case subjects) is taken
to be the "risk haplotype" and carry the feature of interest. If the median
length is equal for the two haplotypes, the mean is used as a tie-breaker. If
the means are also equal, then a warning is generated and by default
haplotype 1 is identified as the risk haplotype.
}
