% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-id_riskhap_inner.r
\name{id_riskhap_inner}
\alias{id_riskhap_inner}
\title{id_riskhap_inner}
\usage{
id_riskhap_inner(ibs, case.ids, site.start, site.end)
}
\arguments{
\item{ibs}{IBS data.frame}

\item{case.ids}{Vector of case identifiers}

\item{site.start}{Genomic position at start of site of interest}

\item{site.end}{Genomic position at end of site of interest}
}
\value{
A vector of risk haplotype IDs
}
\description{
Identify which haplotype is the risk haplotype (at the site of interest)
for each case individual. Called by id.riskhap().
}
\keyword{internal}
