% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b-pairwise_summary.r
\name{pairwise_summary}
\alias{pairwise_summary}
\title{pairwise_summary}
\usage{
pairwise_summary(consensus, id1, id2)
}
\arguments{
\item{consensus}{Data frame with all consensus haplotype segments for each case subject (output from get.consensus)}

\item{id1}{First risk haplotype identifier (should have form like SUBJ1.2 and be an element of the riskhaps vector)}

\item{id2}{Second risk haplotype identifier (should have form like SUBJ3.1 and be an element of the riskhaps vector)}
}
\value{
Start, end, and length of haplotype shared between these subjects
}
\description{
Summarizes overall start and end position of portion of
consensus risk haplotype shared between two particular
case individuals.
}
