% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{riskhaps}
\alias{riskhaps}
\title{Haplotype identifiers for risk haplotypes (i.e., haplotypes carrying
the consensus risk haplotype at the site of interest).}
\format{A vector of length 10, with one haplotype identifier per subject.}
\usage{
riskhaps
}
\description{
Haplotype identifiers for risk haplotypes (i.e., haplotypes carrying
the consensus risk haplotype at the site of interest).
}
\keyword{datasets}
