% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-split_haplotypes.r
\name{split_haplotypes}
\alias{split_haplotypes}
\title{split_haplotypes}
\usage{
split_haplotypes(direction = c("right", "left"), consensus, haps, groups,
  this.hapid, end.pos, splits, ends)
}
\arguments{
\item{direction}{Which direction to move from site of interest. Options are "right" (genomic positions higher than site of interest) and "left" (genomic positions lower than site of interest).}

\item{consensus}{Skeleton data frame to which consensus haplotype segments will be added}

\item{haps}{Matrix of phased haplotypes}

\item{groups}{Internal variable tracking which subjects are in the current haplotype group}

\item{this.hapid}{Consensus haplotype segment currently under consideration}

\item{end.pos}{Current end position of consensus haplotype segment}

\item{splits}{Vector of positions at which consensus haplotype splits into two new segments}

\item{ends}{Vector storing end.pos for all consensus haplotype segments}
}
\value{
Consensus, groups, splits, ends
}
\description{
Splits the current consensus risk haplotype segment into two new groups
}
\keyword{internal}
