% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c-trunc_ibs.r
\name{trunc_ibs_mat}
\alias{trunc_ibs_mat}
\title{trunc_ibs}
\usage{
trunc_ibs_mat(ibs.mat, graphmin = NULL, graphmax = NULL)
}
\arguments{
\item{ibs.mat}{Matrix of pairwise IBS sharing at the site of interest. Likely generated by ibs.matrix()}

\item{graphmin}{Vector of risk haplotype identifiers (one per case subject).}

\item{graphmax}{Genomic position at start of site of interest.}
}
\value{
Matrix of truncated pairwise IBS sharing between case individuals across site of interest.
}
\description{
Truncates the min/max of the IBS matrix to graphmin/graphmax
to make plots easier to read
}
\keyword{internal}
