% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vcf2hap.r
\name{vcf2hap}
\alias{vcf2hap}
\title{vcf2hap}
\usage{
vcf2hap(vcf.file, case.id.file)
}
\arguments{
\item{vcf.file}{Path to the input vcf file}

\item{case.id.file}{Path to the input case ID file}
}
\value{
A matrix of case haplotypes
}
\description{
This function loads a phased vcf file and a case ID file and produes
a haplotype file in required format.
}
\details{
The vcf file should have subject identifiers (column names) that match the
case ID file, its second column should contain the genomic position (in BP
or cM, to match the position of the start and end sites), and the header
row should begin with '#CHROM'. This file may contain extra subjects; the
hap file produced will contain only the cases.

The case ID file should be a text file with a single column containing all
of the case subject identifiers.

The resulting haplotype file has CaseID1.1 and CaseID1.2 as the haplotype
identifiers for a subject with identifier CaseID1.
}
